adapt/fix/move tests + test setup (travis)
