+++ Cargo.toml
@@ -17,6 +17,9 @@ include = [
 [badges]
 travis-ci = { repository = "alteous/gltf" }
 
+[workspace]
+members = ["gltf-derive", "gltf-json", "gltf-importer", "gltf-utils"]
+
 [dependencies]
 gltf-json = { path = "gltf-json", version = "0.9" }
 lazy_static = "0.2"
