+++ gltf-json/tests/minimal_accessor_invalid.gltf
